fix: merge changed files solving the conflicts
