+++ src/app/layout.tsx
@@ -4,13 +4,13 @@ import { Providers } from './_client/providers'
 import MainWrapper from './_components/main-wrapper'
 import { inter } from '@/lib/utils/fonts'
 import { headers } from 'next/headers'
-import InstallPromptDrawer from '@/components/install-prompt-drawer'
+// import InstallPromptDrawer from '@/components/install-prompt-drawer'
 
 export { metadata, viewport } from './_lib/utils/metadata'
 
-type Props = React.PropsWithChildren<LayoutWithSlots<'topbar' | 'bottombar' | 'modal'>>
+type Props = React.PropsWithChildren<LayoutWithSlots<'topbar' | 'bottombar' | 'modal' | 'transactionprogressmodal'>>
 
-export default function RootLayout({ children, bottombar, topbar, modal }: Props) {
+export default function RootLayout({ children, bottombar, topbar, modal, transactionprogressmodal }: Props) {
     const wagmiCookie = headers().get('cookie')
 
     return (
@@ -22,6 +22,7 @@ export default function RootLayout({ children, bottombar, topbar, modal }: Props
                     <MainWrapper>{children}</MainWrapper>
                     {modal}
                     {bottombar}
+                    {transactionprogressmodal}
                     {/* <InstallPromptDrawer /> */}
                 </Providers>
             </body>
